src/main/java/Accessories/DrumSticks.java
package Accessories;

public class DrumSticks implements ISell{

    double buyingPrice;
    double sellingPrice;

    public DrumSticks(double buyingPrice, double sellingPrice) {
        this.buyingPrice = buyingPrice;
        this.sellingPrice = sellingPrice;
    }

    public double getBuyingPrice() {
        return buyingPrice;
    }

    public double getSellingPrice() {
        return sellingPrice;
    }

    public double calculateMarkup() {
        return sellingPrice - buyingPrice;
    }


}
